Syncing updated features

+++ app/Http/Controllers/PostsController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\ValidatePost;
 use App\Post;
 use App\User;
 use Illuminate\Http\Request;
@@ -16,48 +17,36 @@ class PostsController extends Controller
 
     public function index()
     {
-        $ordered = Post::all()->sortByDesc('updated_at');
+        $posts = Post::getPostsInOrder();
         return view('posts.index', [
-            'posts' => $ordered,
+            'posts' => $posts
         ]);
     }
 
-    public function createPost(Request $request)
+    public function createPost()
     {
-//        dd($request);
         return view('posts.create-post');
     }
 
-    public function storePost(Request $request)
+    public function storePost(ValidatePost $request)
     {
-        $this->validate($request, [     // validation
-            'post-title' => 'required|min:10|max:100',
-            'post-text' => 'required|min:10'
-        ]);
         $post = new Post();
         $post->title = $request['post-title'];
         $post->text = $request['post-text'];
-        if ($request->user()->posts()->save($post)) {
-            $message = 'Post was successfully created!';
-        }
-        return redirect()->route('posts.create-post')->with(['message' => $message]);  //redirekto kopā ar mesidžu
+        auth()->user()->posts()->save($post);
+
+        return redirect()->route('posts.create-post')->with(['message' => 'Post was successfully created!']);
     }
 
-//return back()->withInput();
-
-    public function show(int $id)
+    public function show(Post $post, int $id)
     {
-        $post = Post::where('id', $id)->first();
-        $postedBy = $post->where('user_id', $post->user_id)->first()->user;
-        return view('posts.show',
-            ['post' => $post,
-                'postedBy' => $postedBy]);
-
+        $post = $post->find($id);
+        return view('posts.show', ['post' => $post]);
     }
 
-    public function edit(int $id)
+    public function edit(Post $post, int $id)
     {
-        $post = Post::findOrFail($id);
+        $post = $post->find($id);
         if (auth()->user()->id == $post->user_id) {
             return view('posts.edit-post', ['post' => $post]);
         } else {
@@ -65,39 +54,23 @@ class PostsController extends Controller
         }
     }
 
-
-    public function update(Request $request, $id)
+    public function update(Post $post, ValidatePost $request, int $id)
     {
-        $this->validate($request, [
-            'post-title' => 'required|min:10|max:100',
-            'post-text' => 'required|min:10'
-        ]);
-        $post = Post::findOrFail($id);
-        $post->title = $request['post-title'];
-        $post->text = $request['post-text'];
-        $userPosts = $post->user->all()->where('id', $post->user_id);
-//        $post->fill($request->all());
+        $post = $post->find($id);
+        $post->fill(['title' => $request['post-title'], 'text' => $request['post-text']]);
         $post->save();
 
         return redirect()->route('users.posts', [$post->user_id, $post->user->name, $post->user->surname])
-            ->with(['userPosts' => $userPosts,
-                'user' => $post->user, 'message' => 'Your post was successfully updated!']);
+            ->with(['user' => $post->user, 'message' => 'Your post was successfully updated!']);
     }
 
     public function destroy(Post $post)
     {
-        $userPosts = $post->user->all()->where('id', $post->user_id);
-        $deletablePost = $post->where('id', $post->id)->first();
-        $deletablePost->delete();
-
+        if (auth()->user()->id == $post->user_id) {
+            $deletablePost = $post->where('id', $post->id)->first();
+            $deletablePost->delete();
+        }
         return redirect()->route('users.posts', [$post->user_id, $post->user->name, $post->user->surname])
-            ->with(['userPosts' => $userPosts,
-                'user' => $post->user, 'message' => 'Your post was successfully deleted!']);
+            ->with(['user' => $post->user, 'message' => 'Your post was successfully deleted!']);
     }
-
-    public function redirect()
-    {
-        return redirect()->route('users.posts', [auth()->user()->id, auth()->user()->name, auth()->user()->surname]);
-    }
-
 }
